Script to archive raw RAF datasets to the CISL Mass Storage System under the /RAF and /ATD/DATA paths.
Note: This script runs only in the project's "Production/archive" subdirectory under the "dmg" login. It accesses dmg-user environment variables and extracts the project and platform from the path. The complete path MUST follow the pattern /jnet/local/projects/<PROJ>/<PLATFORM>/Production/archive

This script is an attempt to update and consolidate the arch* csh scripts written by Ron Ruth and located in /scr/raf2/Prod_Data/archives/templates.  Impetus for it's development comes from the addition of the HAIS instruments to the data processing stream for HEFT08, ICE-L, PACDEX and subsequent projects.

When this script is complete, it can go under /scr/raf/Prod_Data/archives/templates with the rest of the archive scripts. Actually, once this script starts replacing the template scripts, put all the template scripts into SVN as old versions of this, and put this into SVN and check it out for each project as need it (like all other DMG scripts are handled).



git-svn-id: http://svn.eol.ucar.edu/svn/dmg/tools/mass_store/archAC/trunk@1611 f40853a6-4f38-0410-afbb-e368eb7984ef