This script is designed to be used to install all of the Debian packages required to run on a new viper or vulcan CPU.
A compact flash card needs to be installed and pre-formatted
in the CPU card for this script to completely finish working.

This sript is based upon the guidelines specified in:

  http://wiki.eol.ucar.edu/sew/NIDAS/LinuxOnArcom

This script is almost working...

I am running into issues where the /etc/modprobe.d/ads3
is not present on vulcan after the installation.

There is also an odd dependancy on the missing 'wireless-tools'
package.

svn path=/; revision=4584