Significant refactor:
Recognize GAUS platform and handle separately from aircraft platforms.

Accept messages prefixed with DIGEST and an id, and require the message
digest to match when hashed with the password for the given id.  This way
access can be granted to different sites, ie, gaus vehicles, then later
revoked when not needed.  Currently the secure hash algorithm is RIPEMD160,
implemented with openssl on the server side and with the hashlib module on
the python client side.

Create an execute() method for getting and clearing the sql result.

Don't attempt the execute() if the connection failed.

I have no idea if this still works for the aircraft, but I didn't knowingly
try to break anything.

svn path=/; revision=4880