Change message delimiter of UHSAS from a leading ffff00 to a trailing ffff07.
According to
http://svn.eol.ucar.edu/svn/raf/ManufacturerManuals/sensors/UHSAS_cmds.doc
the data records start with an initial "Tic" sequence (ffff00) and that
"Toc" (ffff01) is no longer used. However on the CU UHSAS during PREDICT,
we saw alternating records beginning with the Toc sequence.  

If NIDAS looks for a trailing ffff07 (end of ADC data) it will work
for units that are sending either Tic or Toc. On both kinds of UHSAS
we start looking for the ffff07 after 220 bytes of data - which
will work on both units which have a different length of A2D housekeeping
array.



svn path=/; revision=1807