Removed postgres database access.
A watched directory will watch for new files and new directories.
Any new directory detected will, in turn, be watched in the same
manor as the the initial directory is.

The reLatest regexp defines a constrained path to prevent falsly
linking to any rogue .jpg(s) that are created in the watched directory.

svn path=/; revision=5409