Added nagios test for missing or broken compact flash cards.  This test pings the DSMs first to verify that they are active before reporting the state of their CF cards.
Cleaned up (reduced) indention after 'if ... return else ...' logic.

This script uses a new init.d script to clean out a temporary file.
That file contains a state variable that stores the length of the previous
days ads3_debug.log file.  That value is used to restrict searches to
the current 'day'.

svn path=/; revision=5576