This is a daemon process that listens for the postgres signal 'killiit'. It is ment to be run on eol-rt-data.fl-ext.ucar.edu.
When this signal is received it kills all postgres process(es) that are
in an '<IDLE> in transaction' state that are accessing the real-time-GV
database.

To signal it simply connect to the exposed postgres database real-time-GV
(on eol-rt-data) and issue a 'NOTIFY killiit;' signal.

To observe the process before and after it is killed use this postgres
command:
SELECT * FROM pg_stat_activity;

Currently it is designed to only deal with hangups on the real-time-GV
database.

svn path=/; revision=6693