Rename raw IR variables from IRT and IRB to IRBV and IRTV per algorithm review committee.  In addition to moving forward, retro-active for NOMADSS and DC3.
IRB and IRT were raw names for old Epply IR.  IRTV and IRBV are for the new Kipp and Zonen.

Note: DC3 default.xml had a PITCH calfile while all the others had PITCH cal in the XML.  ck_xml failed on this file, I repaired to be consistent with other files.


svn path=/; revision=3351