pg_dump will give a different result even with the same database.
Update the script so that a difference check is made on sorted elements of the file.  In this way, if the DB has not changed, but pg_dump gives a different result, it is marked as a non-change.  
Also, if there was a change, inform specifically about the change using the sorted files and leave the new db backup file in place to be checked in.  This will make checkins much easier (since svn diff gives difficult to decipher results owing to pg_dump behavior).
Finally, detect when a previous difference has not been checked in and simply alert the mail recipient of that fact.


svn path=/; revision=7053