Remove CVTAS from <sample> of CVIProcessor.
CVTAS, which is read from the derived data socket,
is still sent to LabView by CVIOutput, as the first field after time.

It has always been a bit of a kludge, since it isn't a <variable>
from a <sensor>. This avoid the warning that SampleAverager prints
saying it can't find a match for the CVTAS variable.

svn path=/; revision=4934