Modify this to only do updates.
This script used to have a set list of directories/repos to clone or update.  i.e. it used to either install or update.

Now it only updates, and there is no fixed list.  If any directory has .svn, then an 'svn update' is performed.  Any directory with a .git directory will perform 'git pull'...and if the .git/config has a submodule entry, 'git submodule update' is performed.

svn path=/; revision=9296