Set ttadjust for the ParoScientific & HoneywellPPT
Engage the time stamp smoothing so that samples fall into individual bins and help eliminate accidental double binning.  This is being engaged for some recent projects.

ttadjust="1 3"

The first value is the number of seconds in a ttadjust period.  After
that amount of time the adjusting will proceed with a new base time
and calculated delta-T

The second value is the number of delta-Ts to consider a data gap.