Remove ttadjust
Seemed to be causing some issues.  In post-processing a two second gap in PSF
was getting smeared over the next five minutes in a sample rate file.

ttadjust was on for Paro & HoneywellPPT

We had a real-time issue, but nclear if it was ttadjust or procSorterLength.
We increased from 1.0 to 1.25 on the plane.