Update build_all
Move from csh to bash.
Defaults to build/install.
Has clean all flag.
Includes most main software for servers.
Nidas/rpms/instrument repos are not included.
Anything that was in the previous commit was still in
svn (ads2 related) after at least 7 years,
so assuming as of 2023 it's not relavant.