Rename CFDC CNTS variable. And fix units.
CNTS conflicts with Butonal CN Counter and drags in a derived concentration we do not want. --> CFCNTS

Units for counters, things that are summed over a second and not averaged, need to be "count", not "counts" or "counts/sec".  Otherwise nimbus will average the data.  We can revisit this if we want to include more units options.